Fix appications api controller and spec path/name
