+++ components/financial_assistance/spec/controllers/financial_assistance/api/v1/applications_controller_spec.rb
@@ -3,20 +3,23 @@ require 'swagger_helper'
 RSpec.describe 'applications', type: :request do
 
   path '/api/v1/applications' do
-    
-    get 'edit an application' do
+
+    get 'retreives all applications' do
       tags Users
       produces 'application/json'
       parameter name: :id, :in => :path, :type => :string 
 
-        response '200' 'index' do
+        response '200', 'index' do
           schema type: :object,
             properties: {
+              id: { type: :string },
               family_id: { type: :string }
             },
             required: [ 'family_id' ]
+          let(:family_id) { BSON::ObjectId.new }
+          let(:application) { FactoryBot.create :financial_assistance_application, :with_applicants, family_id: family_id, aasm_state: 'determined' }
 
-          let(:family_id) { Applications.index(family_id: 'family id') }
+          let(:id) { application.id }
           run_test!
         end
